models/health_predictor.py
import numpy as np
from sklearn.ensemble import RandomForestRegressor
import xgboost as xgb
import config

class HealthPredictor:
    def __init__(self, model_type='random_forest'):
        """
        Initialize the health predictor model.
        
        Args:
            model_type (str): Type of model to use ('random_forest' or 'xgboost')
        """
        self.model_type = model_type
        if model_type == 'random_forest':
            self.model = RandomForestRegressor(**config.MODEL_PARAMS['random_forest'])
        else:
            self.model = xgb.XGBRegressor(**config.MODEL_PARAMS['xgboost'])
            
    def train(self, X, y):
        """
        Train the model on the provided data.
        
        Args:
            X (np.array): Feature matrix
            y (np.array): Target values
        """
        self.model.fit(X, y)
        
    def predict(self, X):
        """
        Make predictions using the trained model.
        
        Args:
            X (np.array): Feature matrix
            
        Returns:
            np.array: Predicted values
        """
        return self.model.predict(X)
        
    def predict_health_impact(self, current_conditions, changes):
        """
        Predict the impact of changes in SDOH factors on health outcomes.
        
        Args:
            current_conditions (dict): Current SDOH conditions
            changes (dict): Proposed changes to SDOH factors
            
        Returns:
            dict: Predicted health outcomes
        """
        # This is a placeholder implementation
        # In a real implementation, this would use the trained model to make predictions
        
        # Example calculation (simplified)
        impact = {
            'life_expectancy': 75 + (changes.get('air_quality', 0) * 0.1) +
                             (changes.get('walkability', 0) * 0.2) +
                             (changes.get('green_space', 0) * 0.15),
            'chronic_disease_risk': 25 - (changes.get('air_quality', 0) * 0.2) -
                                  (changes.get('walkability', 0) * 0.3) -
                                  (changes.get('green_space', 0) * 0.25),
            'healthcare_costs': 100 - (changes.get('air_quality', 0) * 0.3) -
                              (changes.get('walkability', 0) * 0.4) -
                              (changes.get('green_space', 0) * 0.35)
        }
        
        return impact 