requirements.txt
pandas>=1.5.3
numpy>=1.23.5
streamlit>=1.22.0
folium>=0.14.0
requests>=2.28.2
scikit-learn>=1.2.2
xgboost>=1.7.5
torch>=2.0.0
torch-geometric>=2.3.1
shap>=0.41.0
geopandas>=0.12.2
plotly>=5.14.1
python-dotenv>=1.0.0
streamlit-folium>=0.11.0 